make all APIs that take inet port arguments take a guint16

+++ gio/gsocketclient.c
@@ -587,7 +587,7 @@ g_socket_client_connect (GSocketClient       *client,
 GSocketConnection *
 g_socket_client_connect_to_host (GSocketClient  *client,
 				 const gchar    *host_and_port,
-				 gint            default_port,
+				 guint16         default_port,
 				 GCancellable   *cancellable,
 				 GError        **error)
 {
@@ -881,7 +881,7 @@ g_socket_client_connect_async (GSocketClient       *client,
 void
 g_socket_client_connect_to_host_async (GSocketClient        *client,
 				       const gchar          *host_and_port,
-				       gint                  default_port,
+				       guint16               default_port,
 				       GCancellable         *cancellable,
 				       GAsyncReadyCallback   callback,
 				       gpointer              user_data)
