+++ gio/gsocketlistener.c
@@ -324,7 +324,7 @@ g_socket_listener_add_address (GSocketListener  *listener,
  */
 gboolean
 g_socket_listener_add_inet_port (GSocketListener  *listener,
-				 int               port,
+				 guint16           port,
 				 GObject          *source_object,
 				 GError          **error)
 {
